+++ docs/regression_curator/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this gear are documented in this file.
 
 ## 0.1.3
 
+* Updates to consider when a field is in the baseline but not derived vars
 * Fixes bug that crashed on lists of dicts
 
 ## 0.1.2